+++ lib/features/journal/ui/widgets/entry_details/header/entry_detail_header.dart
@@ -1,0 +1,140 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:lotti/classes/journal_entities.dart';
+import 'package:lotti/features/ai/ui/ai_prompt_icon_widget.dart';
+import 'package:lotti/features/journal/state/entry_controller.dart';
+import 'package:lotti/features/journal/ui/widgets/entry_details/header/extended_header_modal.dart';
+import 'package:lotti/features/journal/ui/widgets/entry_details/save_button.dart';
+import 'package:lotti/l10n/app_localizations_context.dart';
+import 'package:lotti/themes/colors.dart';
+import 'package:lotti/themes/theme.dart';
+import 'package:lotti/utils/platform.dart';
+
+class EntryDetailHeader extends ConsumerStatefulWidget {
+  const EntryDetailHeader({
+    required this.entryId,
+    this.inLinkedEntries = false,
+    super.key,
+    this.linkedFromId,
+  });
+
+  final bool inLinkedEntries;
+  final String entryId;
+  final String? linkedFromId;
+
+  @override
+  ConsumerState<EntryDetailHeader> createState() => _EntryDetailHeaderState();
+}
+
+class _EntryDetailHeaderState extends ConsumerState<EntryDetailHeader> {
+  bool showAllIcons = false;
+
+  @override
+  Widget build(BuildContext context) {
+    final provider = entryControllerProvider(id: widget.entryId);
+    final notifier = ref.read(provider.notifier);
+    final entryState = ref.watch(provider).value;
+    if (entryState == null) {
+      return const SizedBox.shrink();
+    }
+
+    final entry = entryState.entry;
+    final id = entryState.entryId;
+
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        Expanded(
+          child: SingleChildScrollView(
+            scrollDirection: Axis.horizontal,
+            child: Row(
+              children: [
+                if (entry is! JournalEvent)
+                  SwitchIconWidget(
+                    tooltip: context.messages.journalFavoriteTooltip,
+                    onPressed: notifier.toggleStarred,
+                    value: entry?.meta.starred ?? false,
+                    icon: Icons.star_outline_rounded,
+                    activeIcon: Icons.star_rounded,
+                    activeColor: starredGold,
+                  ),
+                SwitchIconWidget(
+                  tooltip: context.messages.journalFlaggedTooltip,
+                  onPressed: notifier.toggleFlagged,
+                  value: entry?.meta.flag == EntryFlag.import,
+                  icon: Icons.flag_outlined,
+                  activeIcon: Icons.flag,
+                  activeColor: context.colorScheme.error,
+                ),
+                if (isDesktop)
+                  AiPopUpMenu(
+                    journalEntity: entry,
+                    linkedFromId: widget.linkedFromId,
+                  ),
+                IconButton(
+                  icon: const Icon(Icons.more_horiz),
+                  onPressed: () => ExtendedHeaderModal.show(
+                    context: context,
+                    entryId: id,
+                    inLinkedEntries: widget.inLinkedEntries,
+                    linkedFromId: widget.linkedFromId,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+        if (widget.inLinkedEntries) SaveButton(entryId: widget.entryId),
+      ],
+    );
+  }
+}
+
+class SwitchIconWidget extends StatelessWidget {
+  const SwitchIconWidget({
+    required this.tooltip,
+    required this.onPressed,
+    required this.value,
+    required this.icon,
+    required this.activeIcon,
+    required this.activeColor,
+    super.key,
+  });
+
+  final String tooltip;
+  final void Function() onPressed;
+  final bool value;
+
+  final IconData icon;
+  final IconData activeIcon;
+  final Color activeColor;
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: 40,
+      child: IconButton(
+        splashColor: Colors.transparent,
+        focusColor: Colors.transparent,
+        padding: EdgeInsets.zero,
+        splashRadius: 1,
+        tooltip: tooltip,
+        onPressed: () {
+          if (value) {
+            HapticFeedback.lightImpact();
+          } else {
+            HapticFeedback.heavyImpact();
+          }
+          onPressed();
+        },
+        icon: value
+            ? Icon(
+                activeIcon,
+                color: activeColor,
+              )
+            : Icon(icon),
+      ),
+    );
+  }
+}
